Adding teleport to Dark Tower region
The Dark Tower has long been a black hole of evil... but it also has the
added effect of drawing players in for days until they lose their sanity
(or become a necromancer). A new teleport to Evil Norkos should give our
brave idling heroes a chance to get away from it all, if only to wander
in to more nefarious monsters.