Funhouse: Complying with OSHA
Due to safety violations caused by the Funhouse owners and operators,
the Norkos Council has forced changes to be made to an otherwise
exciting attraction.

Failed teleports in the maze now return the player to before the first
teleport, so should you wisen up and decide you do not, in fact, want to
enter the funhouse, you have the option.
The failed teleport room has been repurposed for the employee break
room.
The supply closet has been labeled properly.